+++ plugins/frontend/backstage-plugin-launchdarkly/package.json
@@ -23,13 +23,14 @@
     "postpack": "backstage-cli package postpack"
   },
   "dependencies": {
+    "@backstage/catalog-model": "^1.6.0",
     "@backstage/core-components": "^0.14.4",
     "@backstage/core-plugin-api": "^1.9.3",
+    "@backstage/plugin-catalog-react": "^1.12.3",
     "@backstage/theme": "^0.5.6",
     "@material-ui/core": "^4.9.13",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.60",
-    "@backstage/plugin-catalog-react": "^1.12.3",
     "lodash": "^4.17.21",
     "react-use": "^17.2.4"
   },
